Move refactored classes to storage module
