+++ eventmesh-storage/eventmesh-storage-standalone/src/main/java/org/apache/eventmesh/storage/standalone/broker/StandaloneBroker.java
@@ -19,6 +19,7 @@ package org.apache.eventmesh.storage.standalone.broker;
 
 import org.apache.eventmesh.storage.standalone.broker.model.MessageEntity;
 import org.apache.eventmesh.storage.standalone.broker.model.TopicMetadata;
+import org.apache.eventmesh.storage.standalone.broker.task.HistoryMessageClear;
 import org.apache.eventmesh.storage.standalone.broker.task.HistoryMessageClearTask;
 
 import org.apache.commons.lang3.tuple.Pair;
@@ -116,7 +117,8 @@ public class StandaloneBroker {
     }
 
     private void startHistoryMessageCleanTask() {
-        Thread thread = new Thread(new HistoryMessageClearTask(messageContainer));
+        HistoryMessageClear historyMessageClear = new HistoryMessageClear(messageContainer);
+        Thread thread = new Thread(new HistoryMessageClearTask(historyMessageClear));
         thread.setDaemon(true);
         thread.setName("StandaloneBroker-HistoryMessageCleanTask");
         thread.start();
